Upgrade to Bootstrap 4.4
This removes all Bootstrap 3 assets, replaces them with just the
minified files of Bootstrap 4.4.1 and updates the views accordingly.

Also this sets the table layout to fixed and takes care of overflowing
headings and table cells with long cassette names.

Some more changes here and there:

- Adds a missing timestamp title attribute.
- Adds a "Back to main app" button to the navbar.
- Removes the navbar toggling for mobile. With just one navbar item,
  this seems not needed and the toggle button was in the way ;)
- No more icon for the Delete button. Glyphicons are not bundled with
  Boostrap anymore.
- Minor cosmetics like making use of muted text and tweaking column
  widths.

Finally, this sets the page container to fluid. This might look a bit
off on huge screens but since the UI is basically just a table, and
cassette names can become pretty long, making as much horizontal space
as possible seems to be appropriate.